Refactor client service and test cases, relocated classes
Deleted the ClientService and ClientServiceI classes as they were no longer needed. Several classes were moved to new directories to better reflect their functionalities. Additionally, new test cases and test object mothers were added for client persistence testing purposes. Updated dependencies after relocation and deletion of classes.
